+++ CHANGELOG.md
@@ -6,10 +6,12 @@
 - Unordered and ordered nested lists are supported.
 - `allowNestedList` option is added.
 - Space-only lines are considered empty.
+- `Element.hasAttribute`, `Element.getAttribute`, `Element.setAttribute` and `Element.removeAttribute` are available.
 
 **⚠ Breaking changes**
 - `allowFootnote` becomes `allowReference`
 - `onFootnote` becomes `onReference`
+- `Element.attr` becomes `Element.attributes`
 - `Element.tagName` returns the tag name in uppercase.